Change EJB interceptor factories to share instances as much as possible
