--- ee/src/main/java/org/jboss/as/ee/component/ConstantInterceptorFactory.java
@@ -1,52 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat Inc., and individual contributors as indicated
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package org.jboss.as.ee.component;
-
-import org.jboss.invocation.Interceptor;
-import org.jboss.invocation.InterceptorContext;
-import org.jboss.invocation.InterceptorFactory;
-import org.jboss.invocation.InterceptorFactoryContext;
-
-/**
- * Interceptor factory that returns a constant value
- *
- * @author Stuart Douglas
- */
-public class ConstantInterceptorFactory implements InterceptorFactory {
-
-    private final Object value;
-
-    public ConstantInterceptorFactory(final Object value) {
-        this.value = value;
-    }
-
-    @Override
-    public Interceptor create(final InterceptorFactoryContext context) {
-        return new Interceptor() {
-
-            @Override
-            public Object processInvocation(final InterceptorContext context) throws Exception {
-                return value;
-            }
-        };
-    }
-}
